Catch-all when no issuer claim is present.
There is nothing in the spec that guarantees an issuer to be found
in a JWT token. I'm quoting from the spec:

These are a set of predefined claims which are not mandatory but recommended, to provide a set of useful, interoperable claims. Some of them are: iss (issuer), exp (expiration time), sub (subject), aud (audience), and others.

I recently experienced that myself while validating Wix
tokens. No issuer.

Since this fine library (thank you!) relies on the issuer to select
configuration options, it can't be used in those cases where no issuer
is present.

This PR proposes a catch-all case to address the problem.

This PR is just a quick hack to get there with minimum fuss. Of course you might want to refine
this.